disastertweetapp/requirements.txt
streamlit==1.29.0
tensorflow==2.13.0
nltk==3.8.1
ntscraper
spacy==3.7.2
st-annotated-text
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1.tar.gz